@@ -52,28 +50,28 @@
     "/node_modules/(?!@redhat-cloud-services)"
   ],
   "dependencies": {
-    "@patternfly/patternfly": "4.171.1",
-    "@patternfly/react-charts": "^6.55.16",
-    "@patternfly/react-core": "^4.239.0",
-    "@patternfly/react-icons": "^4.43.15",
-    "@patternfly/react-styles": "4.42.15",
-    "@patternfly/react-table": "^4.108.0",
-    "@react-pdf/renderer": "^2.1.1",
-    "@redhat-cloud-services/frontend-components": "3.11.10",
-    "@redhat-cloud-services/frontend-components-config": "^4.7.3",
-    "@redhat-cloud-services/frontend-components-notifications": "3.2.5",
-    "@redhat-cloud-services/frontend-components-pdf-generator": "^2.6.12",
-    "@redhat-cloud-services/frontend-components-utilities": "^3.7.6",
+    "@patternfly/patternfly": "^5.2.1",
+    "@patternfly/react-charts": "^7.2.2",
+    "@patternfly/react-core": "^5.2.3",
+    "@patternfly/react-icons": "^5.2.1",
+    "@patternfly/react-styles": "^5.2.1",
+    "@patternfly/react-table": "^5.2.4",
+    "@react-pdf/renderer": "^3.1.14",
+    "@redhat-cloud-services/frontend-components": "^4.2.5",
+    "@redhat-cloud-services/frontend-components-config": "^6.0.11",
+    "@redhat-cloud-services/frontend-components-notifications": "^4.1.0",
+    "@redhat-cloud-services/frontend-components-pdf-generator": "^4.0.5",
+    "@redhat-cloud-services/frontend-components-utilities": "^4.0.7",
     "@unleash/proxy-client-react": "^3.6.0",
     "babel-plugin-transform-imports": "^2.0.0",
     "classnames": "2.2.6",
     "fast-text-encoding": "^1.0.6",
     "moment": "^2.29.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.0",
-    "react-redux": "^7.2.0",
-    "react-router-dom": "^6.8.1",
-    "redux": "^4.0.5",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^7.2.9",
+    "react-router-dom": "^6.21.1",
+    "redux": "^4.2.1",
     "redux-logger": "3.0.6",
     "redux-promise-middleware": "6.1.2"
   },
