@@ -1,27 +1,22 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { EmptyState, EmptyStateBody, EmptyStateIcon, EmptyStateVariant, Title } from '@patternfly/react-core';
+import { EmptyState, EmptyStateBody, EmptyStateIcon, EmptyStateVariant, EmptyStateHeader, EmptyStateFooter  } from '@patternfly/react-core';
 
 export const EmptyStateDisplay = ({ button, color, error, icon, isSmall, text, title }) => (
-    <EmptyState variant={ isSmall ? EmptyStateVariant.small : EmptyStateVariant.large }>
+    <EmptyState variant={ isSmall ? EmptyStateVariant.sm : EmptyStateVariant.lg }>
         { icon && <EmptyStateIcon
             icon={ icon }
             color={ color ? color : null }
             className={ isSmall ? 'small-empty-state-icon' : null }
         /> }
-        <Title
-            headingLevel={ isSmall ? 'h5' : 'h1' }
-            size={ isSmall ? 'md' : 'lg' }
-        >
-            { title }
-        </Title>
+        <EmptyStateHeader titleText={<>{ title }</>} headingLevel={ isSmall ? 'h5' : 'h1' } />
         <EmptyStateBody>
             { text?.join('\n') }
             { error }
-        </EmptyStateBody>
-        { button }
-    </EmptyState>
+        </EmptyStateBody><EmptyStateFooter>
+            { button }
+        </EmptyStateFooter></EmptyState>
 );
 
 EmptyStateDisplay.propTypes = {
