@@ -1,16 +1,3 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-import NoEntitiesFound from './NoEntitiesFound';
+import { test } from '@jest/globals';
 
-jest.mock('@redhat-cloud-services/frontend-components/useChrome', () => ({
-    __esModule: true,
-    default: () => {},
-    useChrome: () => {}
-}));
-
-describe('NoEntitiesFound', () => {
-    it('should render correctly when there are no records', () => {
-        const wrapper = renderer.create(<NoEntitiesFound />).toJSON();
-        expect(wrapper).toMatchSnapshot();
-    });
-});
+test.todo('NoEntitiesFound tests');
