@@ -1,5 +1,5 @@
 import React from 'react';
-import { Tooltip } from '@patternfly/react-core';
+import { Icon, Tooltip } from '@patternfly/react-core';
 import {
     ExclamationCircleIcon, ExclamationTriangleIcon, CheckCircleIcon,
     AngleDoubleUpIcon, AngleDoubleDownIcon,
@@ -11,35 +11,35 @@ import propTypes from 'prop-types';
 const stateDetails = (val) =>  ({
     Oversized: {
         text: 'This system is not fully consuming available resources and may be oversized.',
-        icon: <AngleDoubleUpIcon color='#f09800' size='sm'/>
+        icon: <Icon style={{ color: '#f09800' }} size='md'><AngleDoubleUpIcon /></Icon>
     },
     Undersized: {
         text: 'This system is depleting available resources and may be undersized.',
-        icon: <AngleDoubleDownIcon color='red' size='sm'/>
+        icon: <Icon style={{ color: 'red' }} size='md'><AngleDoubleDownIcon /></Icon>
     },
     Idling: {
         text: 'This system displays higher idling cycles than is expected.',
-        icon: <AutomationIcon size='sm'/>
+        icon: <Icon size='md'><AutomationIcon /></Icon>
     },
     'Storage rightsizing': {
         text: 'This system is experiencing a storage Input/Output bottleneck.',
-        icon: <ExclamationTriangleIcon color='#f09800' size='sm'/>
+        icon: <Icon style={{ color: '#f09800' }} size='md'><ExclamationTriangleIcon /></Icon>
     },
     Optimized: {
         text: 'This system is running efficiently.',
-        icon: <CheckCircleIcon color='green' size='sm'/>
+        icon: <Icon style={{ color: 'green' }} size='md'><CheckCircleIcon /></Icon>
     },
     Crashloop: {
         text: 'This system is experiencing a crash loop.',
-        icon: <ExclamationCircleIcon color='red' size='sm'/>
+        icon: <Icon style={{ color: 'red' }} size='md'><ExclamationCircleIcon /></Icon>
     },
     'Waiting for data': {
         text: 'This system is configured, please allow 24 hours for your upload to complete.',
-        icon: <InProgressIcon color='#2B9AF3' size='sm'/>
+        icon: <Icon style={{ color: '#2B9AF3' }} size='md'><InProgressIcon /></Icon>
     },
     'Under pressure': {
         text: 'System resources adequate but experiencing occasional peaks.',
-        icon: <TachometerAltIcon color='#030303' size='sm'/>
+        icon: <Icon style={{ color: '#030303' }} size='md'><TachometerAltIcon /></Icon>
     }
 }[val] || {});
 
